move: subcomment 폴더 따로 만들어서 정리, styled components 분리
